Add files via upload
Here is discussion for part 2, I am alluding to the two plots for number 2 as figure 2.1 and 2.2 and I am referencing the code for part 2 to be in appendix 2. I would also like to provide error values in this part but I dont know if we have those yet. This is what i have for now.